+++ BACK_END/AquiEstoy MongoDB/Models/FoundPetPostModel.cs
@@ -6,8 +6,8 @@
         public string? NamePet { get; set; }
         public string? Species { get; set; }
         public DateTime? DatePublication { get; set; }
-        public int? Longitud { get; set; }
-        public int? Latitud { get; set; }
+        public double? Longitud { get; set; }
+        public double? Latitud { get; set; }
 
         public string? Email { get; set; }
         public string? Description { get; set; }
@@ -15,3 +15,6 @@
         public string? PersonWhoFound { get; set; }
     }
 }
+
+
+
