+++ BACK_END/AquiEstoy MongoDB/Models/LostPetPostModel.cs
@@ -7,8 +7,8 @@
         public string? NamePet { get; set; }
         public string? Species { get; set; }
         public DateTime? DatePublication { get; set; }
-        public int? Longitud { get; set; }
-        public int? Latitud { get; set; }
+        public double? Longitud { get; set; }
+        public double? Latitud { get; set; }
         public string? Email { get; set; }
         public string? Description { get; set; }
         public int? Reward { get; set; }
@@ -16,3 +16,6 @@
 
     }
 }
+
+
+
